[Fix/#23] ChatView Components 분리 등 코드 정리
