+++ GLINT-iOS/GLINT-iOS/Source/Presentation/Feature/Chat/Components/DateSeparatorView.swift
@@ -7,29 +7,34 @@
 
 import SwiftUI
 
-//struct DateSeparatorView: View {
-//    let date: String
-//    
-//    var body: some View {
-//        HStack {
-//            Rectangle()
-//                .frame(height: 1)
-//                .foregroundColor(.gray.opacity(0.3))
-//            
-//            Text(date)
-//                .font(.system(size: 12))
-//                .foregroundColor(.gray)
-//                .padding(.horizontal, 12)
-//                .padding(.vertical, 4)
-//                .background(Color.white.opacity(0.8))
-//                .clipShape(Capsule())
-//                .lineLimit(1)
-//                .fixedSize(horizontal: true, vertical: false)
-//            
-//            Rectangle()
-//                .frame(height: 1)
-//                .foregroundColor(.gray.opacity(0.3))
-//        }
-//        .padding(.horizontal, 16)
-//    }
-//} 
+// MARK: - Modern Date Separator
+struct DateSeparatorView: View {
+    let date: String
+    
+    var body: some View {
+        HStack {
+            Rectangle()
+                .frame(height: 1)
+                .foregroundStyle(Color.glintTextSecondary.opacity(0.3))
+            
+            Text(date)
+                .font(.pretendardFont(.caption_semi, size: 10))
+                .foregroundStyle(Color.glintTextSecondary)
+                .padding(.horizontal, 12)
+                .padding(.vertical, 6)
+                .background(
+                    Capsule()
+                        .fill(.ultraThinMaterial)
+                        .overlay(
+                            Capsule()
+                                .stroke(Color.glintTextSecondary.opacity(0.2), lineWidth: 1)
+                        )
+                )
+            
+            Rectangle()
+                .frame(height: 1)
+                .foregroundStyle(Color.glintTextSecondary.opacity(0.3))
+        }
+        .padding(.horizontal, 20)
+    }
+}
